Initial commit of working roman numeral converter. We only use the most common roman numerals I, V, X, L, C, D, M and therefore can only convert in the range of 1 - 4999.
All the work is done in the RomanNumeralConverter method which is passed
an arabic numeral. We do a while loop until the arabic number is no
longer greater than zero. Each interation of the loop checks the arabic
number for the greatest possible match of a roman numeral. Once a match
is found the number is appended to the main roman numeral conversion
string and the value of the roman number is subtracted from the main
arabic number. The loop then continues to its next iteration until the
arabic number is completely broken down.

There is also a range validator for good measure.