1. Verify that the existing backend REST endpoints work.
Verified.

I'm not on a windows machine and therefore couldn't use the .bat script.
I installed the gradle ppa here:

https://launchpad.net/~cwchien/+archive/ubuntu/gradle

and was able to build the backend app successfully.

The backend app is running on port 8080 and returns json with an
incrementing id and "Hello World".

2.Create a new REST endpoint at '/convert' that takes an Arabic numeral
request parameter and converts the number to a Roman numeral.

Copied the Greeting controller and class examples as a starting point. I
removed the id counter and changed the name parameter into int number.
Copied the RomanNumeral class from the first test into this project and
imported it into the new Converter class. The converted roman numeral is
now returned in the json content field.

NOTES:

1. We should use dependancy injection and inject the
RomanNumeralConverter class into the Converter class. For this exercise
though, well just initiate the class in the constructor.

2. Better validation and error handiling in the jason service response
would be nice.